Refactor: centralize MP4 box handling
